Rename Java Connector to JavaNetHttp Connector to distinguish from other java connectors (loom, HTTP/3,...) Added documentation
