+++ bom/pom.xml
@@ -80,7 +80,7 @@
             </dependency>
             <dependency>
                 <groupId>org.glassfish.jersey.connectors</groupId>
-                <artifactId>jersey-java-connector</artifactId>
+                <artifactId>jersey-jnh-connector</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
