--- connectors/java-connector/src/main/resources/org/glassfish/jersey/java/connector/localization.properties
@@ -1,21 +1,0 @@
-#
-# Copyright (c) 2021 Oracle and/or its affiliates. All rights reserved.
-#
-# This program and the accompanying materials are made available under the
-# terms of the Eclipse Public License v. 2.0, which is available at
-# http://www.eclipse.org/legal/epl-2.0.
-#
-# This Source Code may also be made available under the following Secondary
-# Licenses when the conditions for such availability set forth in the
-# Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-# version 2 with the GNU Classpath Exception, which is available at
-# https://www.gnu.org/software/classpath/license.html.
-#
-# SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-#
-
-error.body.publisher=Could not determine BodyPublisher for entity.
-error.invalid.class={0} is not an instance of {1}. Ignoring property.
-error.invalid.entity=Could not serialize entity.
-invalid.configurable.component.type=The supplied component "{0}" is not assignable from JerseyClient or JerseyWebTarget.
-expected.connector.provider.not.used=The supplied component is not configured to use a JavaConnectorProvider.
